+++ apis/r/R/write_soma.R
@@ -12,16 +12,16 @@
 #' @param tiledbsoma_ctx Optional \code{\link{SOMATileDBContext}}
 #'
 #' @return The URI to the resulting \code{\link{SOMAExperiment}} generated from
-#' the data contained in \code{x}, returned opened for write
+#' the data contained in \code{x}
 #'
 #' @section Known methods:
 #' \itemize{
 #'  \item \link[tiledbsoma:write_soma.Seurat]{Writing Seurat objects}
+#'  \item \link[tiledbsoma:write_soma.SummarizedExperiment]{Writing SummarizedExperiment objects}
+#'  \item \link[tiledbsoma:write_soma.SingleCellExperiment]{Writing SingleCellExperiment objects}
 #' }
 #'
 #' @export
-#'
-#' @noMd
 #'
 write_soma <- function(x, uri, ..., platform_config = NULL, tiledbsoma_ctx = NULL) {
   UseMethod(generic = 'write_soma', object = x)
@@ -351,7 +351,7 @@ write_soma.TsparseMatrix <- function(
 ) {
   check_package('SeuratObject', version = .MINIMUM_SEURAT_VERSION())
   stopifnot(
-    "'x' must be a data frame" = is.data.frame(x),
+    "'x' must be a data frame" = is.data.frame(x) || inherits(x, 'DataFrame'),
     "'alt' must be a single character value" = is_scalar_character(alt),
     "'axis' must be a single character value" = is_scalar_character(axis),
     "'prefix' must be a single character value" = is_scalar_character(prefix)